Changed pact consumer and provider names to make it clearer that they are for Secom V2
