+++ src/test/java/org/grad/eNav/atonServiceClient/pacts/secomV2S125ServiceClient/S125ServiceClientSecomV2RemoveSubscriptionTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.grad.eNav.atonServiceClient.pacts.secomClientv2;
+package org.grad.eNav.atonServiceClient.pacts.secomV2S125ServiceClient;
 
 import au.com.dius.pact.consumer.MockServer;
 import au.com.dius.pact.consumer.dsl.PactBuilder;
@@ -37,11 +37,11 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
 /**
- * The basic SECOM Client Consumer Remove Subscription Interface Contract
+ * The S125 Aton client Consumer SECOM Remove Subscription Interface Contract
  * Test Class.
  * <p/>
- * This class provides the definition of the consumer-driver contracts for a
- * basic SECOM Client Consumer Remove Subscription interface and generates
+ * This class provides the definition of the consumer-driver contracts for the
+ * S125 AtoN Service Client SECOM Remove Subscription interface and generates
  * the data to be published to the pacts-broker. This can be done through a
  * separate maven goal, so that it doesn't conflict with the development of the
  * service.
@@ -49,8 +49,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
  * @author Nikolaos Vastardis (email: Nikolaos.Vastardis@gla-rad.org)
  */
 @PactConsumerTest
-@PactTestFor(providerName = "SecomService-v2")
-public class BasicSecomV2RemoveSubscriptionTest {
+@PactTestFor(providerName = "SecomV2S125Service")
+public class S125ServiceClientSecomV2RemoveSubscriptionTest {
 
     /**
      * The test object mapper.
@@ -61,7 +61,7 @@ public class BasicSecomV2RemoveSubscriptionTest {
      * SECOM Remove Subscription Pact.
      * @param builder The Pact Builder
      */
-    @Pact(provider="SecomService-v2", consumer="SecomServiceClient-v2")
+    @Pact(provider="SecomV2S125Service", consumer="SecomV2S125ServiceClient")
     public V4Pact removeSubscriptionPact(PactBuilder builder) {
         return builder
                 .given("Test SECOM Remove Subscription Interface")
@@ -83,7 +83,7 @@ public class BasicSecomV2RemoveSubscriptionTest {
      * SECOM Remove Subscription With Bad Body Pact.
      * @param builder The Pact Builder
      */
-    @Pact(provider="SecomService-v2", consumer="SecomServiceClient-v2")
+    @Pact(provider="SecomV2S125Service", consumer="SecomV2S125ServiceClient")
     public V4Pact removeSubscriptionPactWithBadBody(PactBuilder builder) {
         return builder
                 .given("Test SECOM Remove Subscription Interface")
@@ -105,7 +105,7 @@ public class BasicSecomV2RemoveSubscriptionTest {
      * SECOM Remove Subscription With Bad Body Pact.
      * @param builder The Pact Builder
      */
-    @Pact(provider="SecomService-v2", consumer="SecomServiceClient-v2")
+    @Pact(provider="SecomV2S125Service", consumer="SecomV2S125ServiceClient")
     public V4Pact removeSubscriptionPactNotFound(PactBuilder builder) {
         return builder
                 .given("Test SECOM Remove Subscription Interface without subscriptions")
